requirements.txt
streamlit-mic-recorder
InstructorEmbedding
ctransformers
transformers
langchain
streamlit
pypdfium2
librosa
pyaml
chromadb
streamlit_chat
torch
langchain_community